requirements.txt
wandb
gym==0.21.0
gymnasium
stable-baselines3[extra]
python-dotenv
BeautifulSoup4
scikit-learn
scipy 
matplotlib